requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
